pyproject.toml
[project]
name = "shirobase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.8",
    "django-allauth[socialaccount]>=65.13.1",
    "django-auditlog>=3.3.0",
    "django-author>=1.2.0",
    "django-browser-reload>=1.21.0",
    "django-debug-toolbar>=6.1.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-phonenumber-field[phonenumbers]>=8.3.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.2.12",
    "pytest-django>=4.11.1",
    "pytest-sugar>=1.1.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "smartmin>=5.2.2",
    "whitenoise>=6.11.0",
]
